Move createProfile method
