+++ silx/gui/plot/tools/profile/rois.py
@@ -35,7 +35,6 @@ from silx.gui import colors
 
 from silx.gui.plot import items
 from silx.gui.plot.items import roi as roi_items
-from silx.gui.plot.Profile import createProfile
 from . import core
 from silx.gui import utils
 
@@ -132,7 +131,7 @@ class _DefaultImageProfileRoiMixIn(core.ProfileRoiMixIn):
 
         origin = item.getOrigin()
         scale = item.getScale()
-        _coords, _profile, area, _profileName, _xLabel = createProfile(
+        _coords, _profile, area, _profileName, _xLabel = core.createProfile(
             roiInfo=self._getRoiInfo(),
             currentData=currentData,
             origin=origin,
@@ -173,7 +172,7 @@ class _DefaultImageProfileRoiMixIn(core.ProfileRoiMixIn):
         method = self.getProfileMethod()
 
         def createProfile2(currentData):
-            coords, profile, _area, profileName, xLabel = createProfile(
+            coords, profile, _area, profileName, xLabel = core.createProfile(
                 roiInfo=self._getRoiInfo(),
                 currentData=currentData,
                 origin=origin,
@@ -200,7 +199,7 @@ class _DefaultImageProfileRoiMixIn(core.ProfileRoiMixIn):
             _coords, g, _profileName, _xLabel = createProfile2(rgba[..., 1])
             _coords, b, _profileName, _xLabel = createProfile2(rgba[..., 2])
             if rgba.shape[-1] == 4:
-                _coords, a, _profileName, _xLabel = createProfile(rgba[..., 3])
+                _coords, a, _profileName, _xLabel = createProfile2(rgba[..., 3])
             else:
                 a = [None]
             data = core.RgbaProfileData(
@@ -868,7 +867,7 @@ class _DefaultImageStackProfileRoiMixIn(_DefaultImageProfileRoiMixIn):
         assert kind == "2D"
 
         def createProfile2(currentData):
-            coords, profile, _area, profileName, xLabel = createProfile(
+            coords, profile, _area, profileName, xLabel = core.createProfile(
                 roiInfo=self._getRoiInfo(),
                 currentData=currentData,
                 origin=origin,
